Added the possibility to exclude scopes
The definition of the scope was too large for some languages and
as a consequence some keywords were spell checked even though
they were part of the language itself.
With this change it is possible to have a greater control over
which scopes are spell checked.

E.g. In latex "\textit" will not be matched by the spell checker
by specifying "support\.function\..*tex" as one of the
excluded scopes.